+++ packages/portal-core/CHANGELOG.md
@@ -1,5 +1,7 @@
 # @monigo/portal-core
 
+## 0.3.1
+
 ## 0.3.0
 
 ### Minor Changes